♻️ refactor/ZIP-32 : API 통합
